+++ classes_folder/book_folder/book_class.rb
