+++ classes_folder/nameable_class.rb
