+++ classes_folder/person_folder/person_class.rb
