+++ classes_folder/person_folder/student_class.rb
