+++ classes_folder/person_folder/teacher_class.rb
